fix: codefactor adjustments
